Simplify code

+++ .circleci/deploy.sh
@@ -12,24 +12,20 @@ git config --global user.email "cimrroot@gmail.com"
 git config --global user.name "cimrroot"
 git config --global push.default simple
 
-# Install git-lfs
-cd /tmp
-wget https://github.com/git-lfs/git-lfs/releases/download/v2.7.2/git-lfs-linux-386-v2.7.2.tar.gz
-mkdir -p git-lfs
-cd git-lfs
-tar xzf ../git-lfs-linux-386-v2.7.2.tar.gz
-sudo ./install.sh
+cd ~/cimr-d/
+git lfs install
 
-cd ~/cimr-d
-git lfs install 
+# Sync files in "submitted_data" directory to private S3 bucket "cimr-root",
+# then remove submitted_data from git and commit w/o CircleCI.
+if [ -d submitted_data ]; then
+    aws s3 sync submitted_data s3://cimr-root
 
-# Sync files in "submitted_data" directory to private S3 bucket "cimr-root"
-aws s3 sync submitted_data s3://cimr-root
+    git rm -rf submitted_data/*
+    git commit -m "CircleCI: clear submitted_data [skip ci]"
+    git push --force --quiet origin master
+fi
 
 # Sync files in "processed_data" directory to public S3 bucket "cimr-d"
-aws s3 sync processed_data s3://cimr-d
-
-# Remove submitted_data from git and commit w/o CircleCI
-git rm -rf submitted_data/*
-git commit -m "CircleCI: clear submitted_data [skip ci]"
-git push --force --quiet origin master
+if [ -d processed_data ]; then
+    aws s3 sync processed_data s3://cimr-d
+fi
